[MetaSchedule][Refactor] Clarify Integration Logic
